+++ tests/e2e/core-tests/CHANGELOG.md
@@ -11,9 +11,8 @@
 - Added new config variable for Simple Product price to `tests/e2e/env/config/default.json`. Defaults to 9.99
 - Shopper Single Product tests
 - Shopper Checkout Apply Coupon
-
-
 - Shopper Cart Apply Coupon
+- Shopper Variable product info updates on different varations
 
 ## Fixed